fix warning
Former-commit-id: 35bc208cdcf25212290971a020715a70be1abfff

+++ ASJGooglePlaces/Place Photos/ASJPlacePhotos.m
@@ -91,7 +91,7 @@ typedef void (^CallbackBlock)(ASJResponseStatusCode, NSArray *);
 }
 
 - (NSURL *)urlForPlacePhoto:(ASJPhoto *)photo {
-    NSString *stub = [NSString stringWithFormat:@"%@?photoreference=%@&maxwidth=%ld&key=%@", kPlacePhotosSubURL, photo.photoReference, photo.width, [ASJConstants sharedInstance].apiKey];
+    NSString *stub = [NSString stringWithFormat:@"%@?photoreference=%@&maxwidth=%ld&key=%@", kPlacePhotosSubURL, photo.photoReference, (unsigned long)photo.width, [ASJConstants sharedInstance].apiKey];
     stub = [stub stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding];
     NSURL *queryURL = [NSURL URLWithString:stub relativeToURL:self.baseURL];
     return queryURL;
